Add files via upload
Write a loop that tries the following values for max_leaf_nodes from a set of
possible values [5, 25, 50, 100, 250, 500]. Then, call the get_mae function on each value of
max_leaf_nodes. Store the output in some way that allows you to select the value of
max_leaf_nodes that gives the most accurate model on your data. You should return the best
value of max_leaf_nodes and the corresponding MAE value.